edit project folder structure
