+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules/
 coverage/
-dist/main.js
+dist/main.js/
