+++ src/controller/board.js
@@ -1,0 +1,6 @@
+const newBoard = () => {
+  const tiles = [null, null, null, null, null, null, null, null, null];
+  return {
+    tiles
+  }
+};
