+++ src/controller/player.js
@@ -1,0 +1,10 @@
+const newPlayer = (name, symbol, turn) => {
+  const playerName = name;
+  const token = symbol;
+  const playerTurn = turn;
+  return {
+    playerName,
+    token,
+    playerTurn
+  };
+};
