--- src/logic/logic.js
@@ -1,16 +1,0 @@
-const game = () => {
-
-
-
-
-  return {
-    board,
-    moveTile,
-    checkWinCondition,
-    checkDrawCondition,
-    newPlayer,
-    newBoard
-  };
-};
-
-module.exports = game;
